LibWeb: Move stylesheet add/remove invalidation into shared helper
No behavior change.
