+++ Libraries/LibWeb/CSS/StyleSheetInvalidation.h
@@ -80,6 +80,11 @@ WEB_API bool selector_may_match_light_dom_outside_shadow_host(StringView selecto
 // as @property or @keyframes) whose effects are not captured by selector invalidation alone.
 void invalidate_root_for_style_sheet_change(DOM::Node& root, StyleSheetInvalidationSet const&, DOM::StyleInvalidationReason, bool force_broad_invalidation = false);
 
+// Targeted invalidation used after a stylesheet is added to or removed from a Document or ShadowRoot, either via a
+// <style> element or via adoptedStyleSheets. The caller is responsible for updating the sheet's
+// owning_documents_or_shadow_roots and any list-membership before invoking this.
+void invalidate_style_for_stylesheet_change(DOM::Node& document_or_shadow_root, CSSStyleSheet const& sheet, DOM::StyleInvalidationReason);
+
 // Summarize how any currently-active stylesheet in `shadow_root` can escape the shadow subtree. Used by mutation
 // paths that need host-side fallout derived from the whole shadow scope rather than a single sheet.
 ShadowRootStylesheetEffects determine_shadow_root_stylesheet_effects(DOM::ShadowRoot const&);
